Completely rewrite synchronous locking approach
We are now storing both our own "async id" and the managed thread id in
the lock and require both to match for the lock to be obtained.
Obtaining the lock with one unlocked based off the other causes both to
be reset to the current value. When the lock is disposed, they are
restored based off the initial context at the time the lock was called.