Add AsyncLock.TryLock() and AsyncLock.TryLockAsync() methods
These methods attempt to obtain the lock (optionally within a specified
TimeSpan/CancellationToken) and execute a callback with the lock held if
it was obtained successfully. Lock/execution status is returned via a
boolean.

This API protects against misuse by only executing the callback if the
lock was obtained. They do not return an `IDisposable` so we don't run
afoul of C#'s inability to post-facto `use` a variable, e.g.

```csharp
var disposable = @lock.TryLock(0, out var locked);
// How do you automatically dispose `disposable` only if was locked,
// without using a try...catch block?
```

An alternative is to always dispose (i.e. always use `using (...)`) but
sometimes return the actual lock and sometimes a dummy value. This
requires that the user be very careful to always check the `out`
parameter:

```csharp
using (await @lock.TryLockAsync(0, out var locked))
{
    if (locked)
    {
        ....
    }
}
```

(We get around the restriction on async types not having `out`
parameters by using "synchronous async" code that returns a `Task` and
doesn't rely on syntax-to-state-machine conversion.)

But maybe the latter is better?

(Also, the reason for a separate `TimeSpan` overload of `TryLockAsync()`
in addition to the `CancellationToken` flavor is because we can
wait on a `SemaphoreSlim` with a timeout without throwing and catching
any exceptions, whereas a `CancellationToken` approach requires
throwing-and-catching, so this *should* be more performant - although we
are making syscalls to get the current system time as part of the
workaround, so I'm not sure how much better this is without
benchmarking.)